Python_Raspberry/05_Schaltuhren/Schaltuhr_for_multipl_Relais/requirements.txt
Flask>=2.0
RPi.GPIO; platform_system=="Linux"
